Update README.md with notice
There is decent confusion in the community on this project and if it is "dead" or not. Recent interviews with Rui show that it is "dead" and it does not appear to be maintained anymore.

This aims to try and clear up that confusion so they don't install it and wonder why it does not work with recent Xamarin.Forms versions.